+++ main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"github.com/AdconnectDevOps/terraform-provider-pingdom/pingdom"
-	"github.com/hashicorp/terraform-plugin-sdk/plugin"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
 )
 
 func main() {