+++ pingdom/resource_pingdom_team.go
@@ -1,22 +1,24 @@
 package pingdom
 
 import (
+	"context"
 	"fmt"
 	"log"
 	"strconv"
 
-	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/russellcardullo/go-pingdom/pingdom"
 )
 
 func resourcePingdomTeam() *schema.Resource {
 	return &schema.Resource{
-		Create: resourcePingdomTeamCreate,
-		Read:   resourcePingdomTeamRead,
-		Update: resourcePingdomTeamUpdate,
-		Delete: resourcePingdomTeamDelete,
+		CreateContext: resourcePingdomTeamCreate,
+		ReadContext:   resourcePingdomTeamRead,
+		UpdateContext: resourcePingdomTeamUpdate,
+		DeleteContext: resourcePingdomTeamDelete,
 		Importer: &schema.ResourceImporter{
-			State: schema.ImportStatePassthrough,
+			StateContext: schema.ImportStatePassthroughContext,
 		},
 		Schema: map[string]*schema.Schema{
 			"name": {
@@ -47,7 +49,6 @@ func teamForResource(d *schema.ResourceData) (*pingdom.Team, error) {
 		var intSlice []int
 		for i := range interfaceSlice {
 			intSlice = append(intSlice, interfaceSlice[i].(int))
-
 		}
 		team.MemberIDs = intSlice
 	}
@@ -55,30 +56,30 @@ func teamForResource(d *schema.ResourceData) (*pingdom.Team, error) {
 	return &team, nil
 }
 
-func resourcePingdomTeamCreate(d *schema.ResourceData, meta interface{}) error {
+func resourcePingdomTeamCreate(_ context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	client := meta.(*pingdom.Client)
 
 	team, err := teamForResource(d)
 	if err != nil {
-		return err
+		return diag.FromErr(err)
 	}
 
 	log.Printf("[DEBUG] Team create configuration: %#v", d.Get("name"))
 	result, err := client.Teams.Create(team)
 	if err != nil {
-		return err
+		return diag.FromErr(err)
 	}
 
 	d.SetId(strconv.Itoa(result.ID))
 	return nil
 }
 
-func resourcePingdomTeamRead(d *schema.ResourceData, meta interface{}) error {
+func resourcePingdomTeamRead(_ context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	client := meta.(*pingdom.Client)
 
 	id, err := strconv.Atoi(d.Id())
 	if err != nil {
-		return fmt.Errorf("Error retrieving id for resource: %s", err)
+		return diag.FromErr(fmt.Errorf("Error retrieving id for resource: %s", err))
 	}
 	team, err := client.Teams.Read(id)
 	if err != nil {
@@ -86,11 +87,11 @@ func resourcePingdomTeamRead(d *schema.ResourceData, meta interface{}) error {
 			d.SetId("")
 			return nil
 		}
-		return fmt.Errorf("Error retrieving team: %s", err)
+		return diag.FromErr(fmt.Errorf("Error retrieving team: %s", err))
 	}
 
 	if err := d.Set("name", team.Name); err != nil {
-		return err
+		return diag.FromErr(err)
 	}
 
 	memberids := schema.NewSet(
@@ -101,42 +102,42 @@ func resourcePingdomTeamRead(d *schema.ResourceData, meta interface{}) error {
 		memberids.Add(member.ID)
 	}
 	if err := d.Set("member_ids", memberids); err != nil {
-		return err
+		return diag.FromErr(err)
 	}
 
 	return nil
 }
 
-func resourcePingdomTeamUpdate(d *schema.ResourceData, meta interface{}) error {
+func resourcePingdomTeamUpdate(_ context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	client := meta.(*pingdom.Client)
 
 	id, err := strconv.Atoi(d.Id())
 	if err != nil {
-		return fmt.Errorf("Error retrieving id for resource: %s", err)
+		return diag.FromErr(fmt.Errorf("Error retrieving id for resource: %s", err))
 	}
 
 	team, err := teamForResource(d)
 	if err != nil {
-		return err
+		return diag.FromErr(err)
 	}
 
 	log.Printf("[DEBUG] Team update configuration: %#v", d.Get("name"))
 
 	if _, err = client.Teams.Update(id, team); err != nil {
-		return fmt.Errorf("Error updating team: %s", err)
+		return diag.FromErr(fmt.Errorf("Error updating team: %s", err))
 	}
 	return nil
 }
 
-func resourcePingdomTeamDelete(d *schema.ResourceData, meta interface{}) error {
+func resourcePingdomTeamDelete(_ context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	client := meta.(*pingdom.Client)
 
 	id, err := strconv.Atoi(d.Id())
 	if err != nil {
-		return fmt.Errorf("Error retrieving id for resource: %s", err)
+		return diag.FromErr(fmt.Errorf("Error retrieving id for resource: %s", err))
 	}
 	if _, err = client.Teams.Delete(id); err != nil {
-		return fmt.Errorf("Error deleting team: %s", err)
+		return diag.FromErr(fmt.Errorf("Error deleting team: %s", err))
 	}
 
 	return nil